ID-4322 Add correct handling for including tacplus in /etc/nsswitch.conf
We don't want to just overwrite the OS provided /etc/nsswitch.conf since
it might be different between Ubuntu vs. Debian (ONL is Debian based),
or might have other customization.

With these changes we just add tacplus when needed to the file and also
we cleanup if the package is removed.